Development Log
Throughout the week, we worked on finishing the gathering of the data as well as including it in our repository. Most importantly, we were able to organize our projects in an organized manner so we can all start working parallelly on the project. Finally, we were able to compile our project after organizing and creating the script for CMake and make. 

This week, we'll be focusing on parsing and algorithms. Thank you.